Remove the APCu GUI file from the plugin codebase.
This was a requirement for approval of the plugin on moodle.org/plugins.
A download and store mechanism was implemented and runs during installation and upgrade of this plugin. If this mechanism fails, the administrator is informed what he has to do to get the plugin up and running (again).